// frontend/src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      employers: {
        Row: {
          address: string | null;
          company_name: string | null;
          contact_person: string | null;
          description: string | null;
          industry: string | null;
          logo: string | null;
          phone: string | null;
          user_id: string;
        };
        Insert: {
          address?: string | null;
          company_name?: string | null;
          contact_person?: string | null;
          description?: string | null;
          industry?: string | null;
          logo?: string | null;
          phone?: string | null;
          user_id: string;
        };
        Update: {
          address?: string | null;
          company_name?: string | null;
          contact_person?: string | null;
          description?: string | null;
          industry?: string | null;
          logo?: string | null;
          phone?: string | null;
          user_id?: string;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
};